refactor(LoginView): 로그인 화면 리팩토링 및 영상 멈춤 이슈 대응
